Applications: nrf5340_audio: Migration done, to be tested
to be squashed
