+++ applications/nrf5340_audio/src/bluetooth/bt_stream/CMakeLists.txt
@@ -6,7 +6,7 @@
 
 zephyr_library_include_directories(
     broadcast
-    unicast_server
+    unicast
 )
 
 if (CONFIG_BT_BAP_BROADCAST_SOURCE)
@@ -14,8 +14,10 @@ if (CONFIG_BT_BAP_BROADCAST_SOURCE)
 			${CMAKE_CURRENT_SOURCE_DIR}/broadcast/broadcast_source.c)
 endif()
 
-if (CONFIG_TRANSPORT_CIS AND (CONFIG_AUDIO_DEV MATCHES HEADSET))
+
+if (CONFIG_BT_BAP_UNICAST_SERVER)
+	print("JAJAJAJAJAJAJAJAJAJAJAJAJAJAJAJA")
 	target_sources(app PRIVATE
-			${CMAKE_CURRENT_SOURCE_DIR}/unicast_server/unicast_server.c)
+			${CMAKE_CURRENT_SOURCE_DIR}/unicast/unicast_server.c)
 endif()
